backend/requirement.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
distlib==0.3.7
Django==3.2.12
django-appconf==1.0.5
django-cors-headers==4.2.0
django-cryptography==1.1
django-environ==0.10.0
django-filter==23.2
django-resized==1.0.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dynaconf==3.2.2
filelock==3.12.2
hvac==1.1.1
idna==3.4
knox==0.1.14
loguru==0.7.0
numpy==1.25.2
pandas==2.0.3
Pillow==10.0.0
platformdirs==2.6.2
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
validators==0.21.2
virtualenv==20.13.0
win32-setctime==1.1.0